+++ src/payment/checkout/payment-form/utils/form-validators.js
@@ -1,0 +1,78 @@
+import getStates from './countryStatesMap';
+
+// eslint-disable-next-line import/prefer-default-export
+export function getRequiredFields(fieldValues, isBulkOrder) {
+  const {
+    firstName,
+    lastName,
+    address,
+    city,
+    country,
+    state,
+    cardExpirationMonth,
+    cardExpirationYear,
+    organization,
+  } = fieldValues;
+
+  const requiredFields = {
+    firstName,
+    lastName,
+    address,
+    city,
+    country,
+    cardExpirationMonth,
+    cardExpirationYear,
+  };
+
+  if (getStates(country)) {
+    requiredFields.state = state;
+  }
+
+  if (isBulkOrder) {
+    requiredFields.organization = organization;
+  }
+
+  return requiredFields;
+}
+
+export function validateAsciiNames(firstName, lastName) {
+  const errors = {};
+
+  if (
+    firstName
+      && lastName
+      && !/[A-Za-z]/.test(firstName + lastName)
+  ) {
+    errors.firstName = 'payment.form.errors.ascii.name';
+  }
+
+  return errors;
+}
+
+export function validateRequiredFields(values) {
+  const errors = {};
+
+  Object.keys(values).forEach((key) => {
+    if (!values[key]) {
+      errors[key] = 'payment.form.errors.required.field';
+    }
+  });
+
+  return errors;
+}
+
+export function validateCardDetails(cardExpirationMonth, cardExpirationYear) {
+  const errors = {};
+
+  const currentMonth = new Date().getMonth() + 1;
+  const currentYear = new Date().getFullYear();
+  if (
+    cardExpirationMonth
+      && parseInt(cardExpirationMonth, 10) < currentMonth
+      && parseInt(cardExpirationYear, 10) === currentYear
+  ) {
+    errors.cardExpirationMonth = 'payment.form.errors.card.expired';
+  }
+
+  return errors;
+}
